backend/requirements.txt
asgiref==3.8.1
bidict==0.23.1
dj-database-url==2.2.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
h11==0.14.0
Markdown==3.7
mysqlclient==2.2.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
simple-websocket==1.0.0
sqlparse==0.5.1
typing_extensions==4.12.2
wsproto==1.2.0
